requirements.txt
flask
gunicorn
flask-cors
python-dotenv
redis
fakeredis
pytest

